Resolve errors
